renderer: standardise function naming for rendering and improve reusability
